+++ src/get.js
@@ -2,16 +2,17 @@ import curry from './curry'
 import get from './uncurried/get'
 
 /**
- * Gets a property of an object.
+ * Gets a property of an object using a key path.
  *
  * @function
- * @param {String} k The key to get.
+ * @param {Array|String} ks The key path.
  * @param {Object} o The object.
- * @returns The property at the key `k` in the object `o`.
+ * @returns The property at the key path `ks` in the object `o`.
  * @example
  *
  * import { get } from 'fkit'
- * const person = { name: 'Jane', age: 20, city: 'Melbourne' }
+ * const person = { name: 'Jane', age: 20, address: { city: 'Melbourne', country: 'Australia' } }
  * get('name', person) // 'Jane'
+ * get('address.city', person) // 'Melbourne'
  */
 export default curry(get)