+++ src/permutations.js
@@ -7,6 +7,37 @@ import id from './id'
 import prepend from './prepend'
 import tail from './tail'
 
+function permutations_ (bs_, cs) {
+  const b = head(bs_)
+  const bs = tail(bs_)
+
+  return empty(bs_) ? []
+    : foldRight(
+      interleave,
+      permutations_(bs, prepend(b, cs)),
+      permutations(cs)
+    )
+
+  function interleave (ds, r) {
+    return interleave_(id, ds)[1]
+
+    function interleave_ (f, es_) {
+      if (empty(es_)) {
+        return [bs, r]
+      } else {
+        const e = head(es_)
+        const es = tail(es_)
+        const s = interleave_(compose(f, prepend(e)), es)
+
+        return [
+          prepend(e, s[0]),
+          prepend(f(concat(b, e, s[0])), s[1])
+        ]
+      }
+    }
+  }
+}
+
 /**
  * Calculates the permutations of a list.
  *
@@ -20,36 +51,5 @@ import tail from './tail'
  * permutations('abc') // ['abc', 'bac', 'cba', 'bca', 'cab', 'acb']
  */
 export default function permutations (as) {
-  const permutations_ = (bs_, cs) => {
-    const b = head(bs_)
-    const bs = tail(bs_)
-
-    return empty(bs_) ? []
-      : foldRight(
-        interleave,
-        permutations_(bs, prepend(b, cs)),
-        permutations(cs)
-      )
-
-    function interleave (ds, r) {
-      return interleave_(id, ds)[1]
-
-      function interleave_ (f, es_) {
-        if (empty(es_)) {
-          return [bs, r]
-        } else {
-          const e = head(es_)
-          const es = tail(es_)
-          const s = interleave_(compose(f, prepend(e)), es)
-
-          return [
-            prepend(e, s[0]),
-            prepend(f(concat(b, e, s[0])), s[1])
-          ]
-        }
-      }
-    }
-  }
-
   return prepend(as, permutations_(as, []))
 }