+++ src/uncurried/get.js
@@ -1,3 +1,6 @@
-export default function get (k, o) {
-  return o[k]
+import isString from '../internal/isString'
+
+export default function get (ks, o) {
+  ks = isString(ks) ? ks.split('.') : ks
+  return ks.reduce((a, k) => (a !== undefined) ? a[k] : undefined, o)
 }