+++ src/update.test.js
@@ -1,9 +1,19 @@
+import always from './always'
+import inc from './inc'
 import update from './update'
 
 describe('update', () => {
   it('updates the given property', () => {
-    const target = { name: 'Jane', age: 20 }
-    const result = update('age')(a => a + 1)(target)
-    expect(result).toHaveProperty('age', 21)
+    const target = {
+      name: 'Jane',
+      age: 20,
+      address: {
+        city: 'Melbourne',
+        country: 'Australia'
+      }
+    }
+
+    expect(update('age')(inc)(target)).toHaveProperty('age', 21)
+    expect(update('address.city')(always('Sydney'))(target)).toHaveProperty('address.city', 'Sydney')
   })
 })